Refactor Grape Resource helpers, add clean_params helper method
