+++ app/api/v1/helpers/api_helper.rb
@@ -2,86 +2,68 @@ module API
   module V1
     module Helpers
       module APIHelper
-        extend ActiveSupport::Concern
+        def logger
+          API.logger
+        end
 
-        included do
-          def logger
-            API.logger
+        # Authentication and Authorization
+        def authenticate!
+          unauthorized! unless current_user
+        end
+
+        def authorize!(action, subject)
+          unless abilities.allowed?(current_user, action, subject)
+            forbidden!
           end
+        end
 
-          # Authentication and Authorization
-          def authenticate!
-            unauthorized! unless current_user
+        def require_scope!(scopes)
+          return unless find_access_token
+          scopes = [scopes] unless scopes.kind_of? Array
+
+          unless (find_access_token.scopes.to_a & scopes) == scopes
+            # TODO: Scopes are historically completely broken in Cortex. This is quite the security issue: fix!
+            puts 'SCOPES are currently being IGNORED'
+            # forbidden!
           end
+        end
 
-          def authorize!(action, subject)
-            unless abilities.allowed?(current_user, action, subject)
-              forbidden!
-            end
-          end
+        def current_tenant
+          current_user.tenant
+        end
 
-          def require_scope!(scopes)
-            return unless find_access_token
-            scopes = [scopes] unless scopes.kind_of? Array
+        def can?(object, action, subject)
+          abilities.allowed?(object, action, subject)
+        end
 
-            unless (find_access_token.scopes.to_a & scopes) == scopes
-              # TODO: Scopes are historically completely broken in Cortex. This is quite the security issue: fix!
-              puts 'SCOPES are currently being IGNORED'
-              # forbidden!
-            end
-          end
+        # API Errors
+        def bad_request!
+          render_api_error!('(400) Bad Request', 400)
+        end
 
-          def current_tenant
-            current_user.tenant
-          end
+        def forbidden!
+          render_api_error!('(403) Forbidden', 403)
+        end
 
-          def can?(object, action, subject)
-            abilities.allowed?(object, action, subject)
-          end
+        def not_found!
+          render_api_error!('(404) Not found', 404)
+        end
 
-          # API Errors
-          def bad_request!
-            render_api_error!('(400) Bad Request', 400)
-          end
+        def unauthorized!
+          render_api_error!('(401) Unauthorized', 401)
+        end
 
-          def forbidden!
-            render_api_error!('(403) Forbidden', 403)
-          end
+        def render_api_error!(message, status)
+          error!({message: message}, status)
+        end
 
-          def not_found!
-            render_api_error!('(404) Not found', 404)
-          end
+        private
 
-          def unauthorized!
-            render_api_error!('(401) Unauthorized', 401)
-          end
-
-          def render_api_error!(message, status)
-            error!({message: message}, status)
-          end
-
-          private
-
-          def abilities
-            @abilities ||= begin
-              abilities = Six.new
-              abilities << Abilities::Ability
-              abilities
-            end
-          end
-
-          def remove_params(params, *remove)
-            new_params = params
-            remove.each do |r|
-              new_params.delete(r)
-            end
-            new_params
-          end
-
-          def clean_params(params)
-            Hashie::Mash.new(params.reject do |_,v|
-              v.blank?
-            end)
+        def abilities
+          @abilities ||= begin
+            abilities = Six.new
+            abilities << Abilities::Ability
+            abilities
           end
         end
       end
