+++ app/api/v1/helpers/media_helper.rb
@@ -9,10 +9,6 @@ module API
         def media!
           media || not_found!
         end
-
-        def media_params
-          clean_params(params)
-        end
       end
     end
   end
