+++ app/api/v1/helpers/users_helper.rb
@@ -9,10 +9,6 @@ module API
         def user!
           user || not_found!
         end
-
-        def user_params
-          clean_params(params)
-        end
       end
     end
   end
