+++ app/api/v1/resources/bulk_jobs.rb
@@ -4,8 +4,6 @@ module API
   module V1
     module Resources
       class BulkJobs < Grape::API
-        helpers Helpers::ParamsHelper
-
         resource :bulk_jobs do
           include Grape::Kaminari
           helpers Helpers::BulkJobsHelper
