+++ app/api/v1/resources/categories.rb
@@ -5,7 +5,6 @@ module API
     module Resources
       class Categories < Grape::API
         resource :categories do
-
           desc 'Show all categories', { entity: Entities::Category, nickname: "showAllCategories" }
           params do
             optional :depth, default: 1, type: Integer, desc: "Minimum category depth"
