+++ app/api/v1/resources/occupations.rb
@@ -4,9 +4,7 @@ module API
   module V1
     module Resources
       class Occupations < Grape::API
-
         resource :occupations do
-
           desc 'Show all occupations', { entity: Entities::Occupation, nickname: "showAllOccupations" }
           get do
             present ::Onet::Occupation.all, with: Entities::Occupation
