+++ app/api/v1/resources/snippets.rb
@@ -4,8 +4,6 @@ module API
   module V1
     module Resources
       class Snippets < Grape::API
-        helpers Helpers::ParamsHelper
-
         resource :snippets do
           include Grape::Kaminari
           helpers Helpers::SnippetsHelper
